Move retrieveItemListFromCSV logic to PosCsvUtils
