+++ lib/bloc/pos_catalog/pos_csv_exporter.dart
@@ -1,14 +1,16 @@
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:breez/bloc/pos_catalog/model.dart';
 import 'package:breez/logger.dart';
 import 'package:csv/csv.dart';
+import 'package:flutter/foundation.dart';
 import 'package:path_provider/path_provider.dart';
 
 class PosCsvUtils {
-  final List paymentList;
+  final List itemList;
 
-  PosCsvUtils(this.paymentList);
+  PosCsvUtils({this.itemList});
 
   Future export() async {
     log.info("export pos items started");
@@ -21,9 +23,9 @@ class PosCsvUtils {
   List _generateList() {
     log.info("generating payment list started");
     List<List<dynamic>> paymentList =
-        List.generate(this.paymentList.length, (index) {
+        List.generate(this.itemList.length, (index) {
       List paymentItem = List();
-      Item paymentInfo = this.paymentList.elementAt(index);
+      Item paymentInfo = this.itemList.elementAt(index);
       paymentItem.add(paymentInfo.id.toString());
       paymentItem.add(paymentInfo.name);
       paymentItem.add(paymentInfo.sku);
@@ -60,5 +62,47 @@ class PosCsvUtils {
     filePath += ".csv";
     log.info("create breez pos items path finished");
     return filePath;
+  }
+
+  Future retrieveItemListFromCSV(File csvFile) async {
+    log.info("retrieve item list from csv started");
+    List csvList = await csvFile
+        .openRead()
+        .transform(utf8.decoder)
+        .transform(new CsvToListConverter())
+        .toList();
+    log.info("header control started");
+    List<String> headerRow = List<String>.from(csvList.elementAt(0));
+    var defaultHeaders = [
+      "ID",
+      "Name",
+      "SKU",
+      "Image URL",
+      "Currency",
+      "Price",
+    ];
+    // Need a more sophisticated control here. Check #1
+    if (listEquals(headerRow, defaultHeaders)) {
+      throw Exception("INVALID_FILE");
+    }
+    // remove header row
+    csvList.removeAt(0);
+    log.info("header control finished");
+    // create items list
+    var itemsList = <Item>[];
+    csvList.forEach((csvItem) {
+      // #1: We should extend this so our users will be able
+      // to import files that does not have this exact column order.
+      Item item = Item(
+          id: csvItem[0],
+          name: csvItem[1],
+          sku: csvItem[2].toString(),
+          imageURL: csvItem[3] != "null" ? csvItem[3] : null,
+          currency: csvItem[4] != "null" ? csvItem[4] : null,
+          price: csvItem[5]);
+      itemsList.add(item);
+    });
+    log.info("retrieving item list from csv finished");
+    return itemsList;
   }
 }
